notatrix: Add type annotations for all "get-loss" calculations
