+++ notatrix/src/formats/notatrix-serial/get-loss.ts
@@ -1,0 +1,6 @@
+import type {Sentence} from "../../nx/sentence";
+
+export function getLoss(sent: Sentence): string[] {
+  // do nothing, can't lose info on this one
+  return [];
+}
