+++ notatrix/src/formats/plain-text/generator.js
@@ -4,7 +4,7 @@ const _ = require("underscore");
 
 const utils = require("../../utils");
 const GeneratorError = utils.GeneratorError;
-const getLoss = require("./get-loss")
+const getLoss = require("./get-loss").getLoss;
 
 module.exports = (sent, options) => {
   if (!sent.isParsed)
