+++ notatrix/src/formats/sd/get-loss.ts
@@ -1,19 +1,18 @@
-"use strict";
+import * as _ from "underscore";
 
-const _ = require("underscore");
+import {FIELDS} from "./fields";
+import type {Sentence} from "../../nx/sentence";
+import type {TokenSerial} from "../../nx/base-token";
 
-const utils = require("../../utils");
-const fields = require("./fields");
-
-module.exports = sent => {
+export function getLoss(sent: Sentence): string[] {
   const serial = sent.serialize();
-  let losses = new Set();
+  let losses: Set<string> = new Set();
 
   serial.tokens.forEach(token => {
     if (token.heads && token.heads.length > 1)
       losses.add("enhanced dependencies");
 
-    Object.keys(_.omit(token, fields.FIELDS)).forEach(field => {
+    Object.keys(_.omit(token, FIELDS)).forEach(field => {
       switch (field) {
       case ("uuid"):
       case ("index"):
@@ -32,11 +31,11 @@ module.exports = sent => {
         break;
 
       default:
-        if (token[field])
+        if (token[field as keyof TokenSerial])
           losses.add(field);
       }
     })
   });
 
   return Array.from(losses);
-};
+}
